requirements.txt
transformers
sentence-transformers
opencv-python
pytesseract
pdfplumber
fpdf
scikit-learn
Pillow
matplotlib
